requirements.txt
asyncer==0.0.2
fastapi==0.88.0
requests==2.28.*
pyhumps==3.8.0
pyjwt==2.6.0
strivelogger==1.0.0
uvicorn==0.20.0